# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Protocolo Cirúrgico Ortognático — Clínica da Face
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /phrases:
    get:
      operationId: listPhrases
      tags: [phrases]
      summary: List clinical phrases
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Phrases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phrase"
    post:
      operationId: createPhrase
      tags: [phrases]
      summary: Create a clinical phrase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhraseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phrase"
...
  /phrases/{id}:
    patch:
      operationId: updatePhrase
      tags: [phrases]
      summary: Update a clinical phrase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhraseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phrase"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePhrase
      tags: [phrases]
      summary: Delete a clinical phrase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /protocols/{id}/files-3d:
    get:
      operationId: listFiles3d
      tags: [files3d]
      summary: List 3D files for a protocol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 3D files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File3d"
        "404":
          description: Protocol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createFile3d
      tags: [files3d]
      summary: Register a 3D file after upload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/File3dInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File3d"
        "404":
          description: Protocol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /protocols/{id}/files-3d/reorder:
    post:
      operationId: reorderFiles3d
      tags: [files3d]
      summary: Reorder 3D files
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File3d"
  /protocols/{id}/files-3d/{fileId}:
    patch:
      operationId: updateFile3d
      tags: [files3d]
      summary: Update 3D file metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/File3dUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File3d"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteFile3d
      tags: [files3d]
      summary: Delete a 3D file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Protocol schemas ───────────────────────────────────────
...
    Phrase:
      type: object
      required: [id, category, text, createdAt, updatedAt]
      properties:
        id: { type: integer }
        category: { type: string }
        subcategory: { type: string }
        text: { type: string }
        isCustom: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }

    # ─── Storage schemas ────────────────────────────────────────
...
    File3d:
      type: object
      required: [id, protocolId, objectPath, servingUrl, fileFormat, fileType, displayOrder, createdAt, updatedAt]
      properties:
        id: { type: integer }
        protocolId: { type: integer }
        objectPath: { type: string }
        servingUrl: { type: string }
        originalName: { type: ["string", "null"] }
        fileFormat:
          type: string
          enum: [stl, ply]
        fileType:
          type: string
          enum: [scanner_intraoral, modelo_dentario, maxila, mandibula, cranio, splint, guia_cirurgica, outro]
        origin: { type: ["string", "null"] }
        fileDate: { type: ["string", "null"] }
        version: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        displayOrder: { type: integer }
        includeInPdf: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    File3dInput:
      type: object
      required: [objectPath, servingUrl, fileFormat, fileType]
      properties:
        objectPath: { type: string }
        servingUrl: { type: string }
        originalName: { type: string }
        fileFormat:
          type: string
          enum: [stl, ply]
        fileType:
          type: string
          enum: [scanner_intraoral, modelo_dentario, maxila, mandibula, cranio, splint, guia_cirurgica, outro]
        origin: { type: string }
        fileDate: { type: string }
        version: { type: string }
        notes: { type: string }
        displayOrder: { type: integer }
        includeInPdf: { type: boolean }
    File3dUpdate:
      type: object
      properties:
        fileType:
          type: string
          enum: [scanner_intraoral, modelo_dentario, maxila, mandibula, cranio, splint, guia_cirurgica, outro]
        origin: { type: string }
        fileDate: { type: string }
        version: { type: string }
        notes: { type: string }
        displayOrder: { type: integer }
        includeInPdf: { type: boolean }

    # ─── Plate catalogue schemas ─────────────────────────────────
...
    ReorderInput:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items:
            type: integer
    PhraseInput:
      type: object
      required: [category, text]
      properties:
        category: { type: string }
        subcategory: { type: string }
        text: { type: string }
        isCustom: { type: boolean }
    PhraseUpdate:
      type: object
      properties:
        category: { type: string }
        subcategory: { type: string }
        text: { type: string }